refactor(storage): move framed log I/O out of domain
